名前変更(pyopentd.py to mein.py)、Caseクラスの作成
